Update helper.h
Helper splits input strings and splits out information.It also generates appropriate components and add them to their respective vector.I tried to make it as less rigid as possible. But I don't thinks it's even close. That means you have to change a lot of the code to add a single element. Well, that sucks. I will be documenting where & how to make changes to the code to add new elements. Considering it as a simple simulator, we won't be needing those too much. I think...